pyproject.toml
[project]
name = "travel-agent-runtime-demo"
version = "0.1.0"
description = "Simple AgentCore Runtime demo using a travel planning agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Runtime deps (also in requirements.txt for the container)
    "bedrock-agentcore>=0.1.0",
    "strands-agents>=1.0.0",
    "boto3>=1.39.7",
    "botocore>=1.39.7",
    # Deploy tooling (local only - NOT in requirements.txt)
    "bedrock-agentcore-starter-toolkit>=0.1.0",
    # Script tooling
    "click>=8.2.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "test"]
